cloudrun/requirements.txt
Flask==3.0.3
gunicorn==22.0.0

google-api-python-client==2.141.0
google-cloud-documentai==3.8.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==6.1.2

notion-client==2.2.1

opencv-python-headless==4.10.0.84
numpy==2.0.1
Pillow==10.4.0
